Fix range gotcha getting a pointer to an element
The address of the element assigned in the range will be the
same in all the iterations of the array, so using a pointer to it inside
the loop will make the pointed value to be the last assigned value in
all the iterations.

Creating a copy of the element inside the array fixes the issue as now
we have a new structure in every iteration.